Added another command line argument
Added another command line argument to specify the name of the new app being
deployed. Previously the name of the new app would just append a new in front
of the new app which looked bad for multiple safe-scale calls. This gives the
user more control and also solves the issue with duplicate names. Also fixed
tests to abide by the change in arguments.